ETH Plugin (#265)
* Initial implementation
* Eth plugin update
* Eth plugin update
* WIP
* Eth wip
* added payoutDetectionType config
* payoutDetectionType wip
* msg tweaks
* requested changes
* allow the rpc version specification at the plugin

+++ altchain-sdk/src/main/kotlin/org/veriblock/sdk/alt/model/SecurityInheritingBlock.kt
@@ -6,7 +6,7 @@ data class SecurityInheritingBlock(
     val previousHash: String,
     val confirmations: Int,
     val version: Short,
-    val nonce: Int,
+    val nonce: Long,
     val merkleRoot: String,
     val difficulty: Double,
     val coinbaseTransactionId: String,
